requirements.txt
flask
flask-cors
python-dotenv
PyPDF2
google-generativeai
gunicorn
